# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Wallet API Documentation",
        "description": "Hệ thống REST API Ví Điện Tử (E-Wallet) - Laravel 11",
        "contact": {
            "email": "admin@e-wallet.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "E-Wallet API Server"
        }
    ],
    "paths": {
        "/api/v1/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Danh sách địa chỉ của người dùng",
                "operationId": "getAddresses",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Thêm địa chỉ mới",
                "operationId": "storeAddress",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "street"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "home"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "123 Nguyễn Huệ"
                                    },
                                    "ward": {
                                        "type": "string",
                                        "example": "Bến Nghé"
                                    },
                                    "district": {
                                        "type": "string",
                                        "example": "Quận 1"
                                    },
                                    "province": {
                                        "type": "string",
                                        "example": "Hồ Chí Minh"
                                    },
                                    "is_default": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đã tạo mới địa chỉ thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/addresses/{id}": {
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Cập nhật thông tin địa chỉ",
                "operationId": "updateAddress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "street": {
                                        "type": "string",
                                        "example": "145 Nguyễn Huệ"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Xóa địa chỉ",
                "operationId": "deleteAddress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}